Tags pull from firestore.
