--- app/src/main/java/com/example/idea/Types/RoomTagDescriptor.java
@@ -1,33 +1,0 @@
-package com.example.idea.Types;
-
-import android.support.annotation.StringDef;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-public class RoomTagDescriptor {
-
-    private final String roomTag;
-
-    // Describes when annotation is to be discarded
-    @Retention(RetentionPolicy.SOURCE)
-
-    // Enumerating valid values for this interface
-    @StringDef({TAG_KITCHEN, TAG_LIVINGROOM, TAG_BEDROOM, TAG_PATIO, TAG_BATHROOM})
-
-    // Interface for validating String types
-    public @interface RoomTagDef{}
-
-    // The tags declared as constants
-    public static final String TAG_KITCHEN = "kitchen";
-    public static final String TAG_LIVINGROOM = "living room";
-    public static final String TAG_BEDROOM = "bedroom";
-    public static final String TAG_PATIO = "patio";
-    public static final String TAG_BATHROOM = "bathroom";
-
-    // Marks the arg as restricted to the enumerated types
-    public RoomTagDescriptor(@RoomTagDef String roomTag) {
-        this.roomTag = roomTag;
-    }
-
-}
